requirements.txt
certifi==2019.11.28
cycler==0.10.0
diffimg==0.2.3
future==0.18.2
idna==2.9
kiwisolver==1.2.0
matplotlib==3.2.1
pandas==1.0.3
Pillow==7.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
torch==1.5.0
torchvision==0.6.0
tqdm==4.45.0